move ci testing to dev_tools
