+++ crates/bevy_internal/Cargo.toml
@@ -113,12 +113,7 @@ webgpu = [
 ]
 
 # enable systems that allow for automated testing on CI
-bevy_ci_testing = [
-  "bevy_dev_tools/bevy_ci_testing",
-  "bevy_time/bevy_ci_testing",
-  "bevy_render?/bevy_ci_testing",
-  "bevy_render?/ci_limits",
-]
+bevy_ci_testing = ["bevy_dev_tools/bevy_ci_testing", "bevy_render?/ci_limits"]
 
 # Enable animation support, and glTF animation loading
 animation = ["bevy_animation", "bevy_gltf?/bevy_animation"]
